삭제 처리

+++ team_b/src/main/java/kr/ac/ync/domain/GameInfoVO.java
@@ -5,7 +5,7 @@ import lombok.Data;
 
 @Data
 public class GameInfoVO {
-	private int game_num ;
+	private Long game_num;
 	private String title;
 	private String info;
 	private String korean;
